requirements.txt
Flask==3.0.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
tinydb==4.8.0